infrastructure/requirements.txt
# AWS CDK Infrastructure Dependencies
# Generated for Python 3.12

aws-cdk-lib>=2.170.0
constructs>=10.0.0
boto3>=1.35.0
